Add files via upload
The screen shots are pictures of the parse dashboard, showing the objects created by the user and and the new user object created when "bob" signed up. I attempted to make a gif with these screen shots as uploads to my app, but the resolution is terrible and unreadable. So to provide a clear picture and proof that the app works as intended, I have provided these for your viewing enjoyment!